Extracted components/functions from Trainer.js to improve code organisation
